Sorting of priority done
