+++ src/Components/TodoList/ListItem/TodoListItemElement/TodoListItemElement.tsx
@@ -5,10 +5,10 @@ import TodoDeleteButton from '../../../TodoDeleteButton/TodoDeleteButton';
 import TodoDoneToggle from '../../../TodoDoneToggle/TodoDoneToggle';
 import useTodoListStore from '../../../../Hooks/useTodoListStore';
 import { TodoListItem } from '../../../../Stores/TodoListStore';
-import useTodoListItemColor from '../../useTodoListItemColor';
+import TodoListItemColor from '../../TodoListItemColor';
 
 const TodoListItemElement: FC<TodoListItem> = ({
-    id, text, isDone,
+    id, text, isDone, priority,
 }) => {
     const todoListStore = useTodoListStore();
     const handleOnDoubleClick = (): void => {
@@ -24,7 +24,7 @@ const TodoListItemElement: FC<TodoListItem> = ({
             <Grid item xs={9}>
                 <Typography
                     variant="h4"
-                    sx={{ textDecoration: isDone ? 'line-through' : 'none', color: useTodoListItemColor(id) }}
+                    sx={{ textDecoration: isDone ? 'line-through' : 'none', color: TodoListItemColor(priority) }}
                     onDoubleClick={handleOnDoubleClick}
                 >
                     {text}
