+++ src/Components/TodosCardList/TodosCardItem/TodosCardItemList/useItemsListToRender.ts
@@ -1,7 +1,7 @@
 import { AMOUNT } from '../../../../constants';
 import useTodoListStore from '../../../../Hooks/useTodoListStore';
 import { TodoListItem } from '../../../../Stores/TodoListStore';
-import useItemsToRenderPriority from '../../../TodoList/useItemsToRenderPriority';
+import useItemsToRenderPriority from '../../../TodoList/TodoListItemsPriority';
 
 const prepareCardList = (currentTodoCardList:TodoListItem[]): TodoListItem[] => currentTodoCardList
     .slice(0, AMOUNT);
